chore: nicer reset button
