+++ src/client/hooks/useReset/userReset.module.scss
@@ -1,0 +1,3 @@
+.snackbar {
+    max-width: rem(600px);
+}
